+++ src/main/java/com/coolerpromc/uncrafteverything/blockentity/custom/AbstractUncraftingTableBE.java
@@ -31,10 +31,7 @@ import net.minecraftforge.network.PacketDistributor;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
+import java.util.*;
 import java.util.regex.Pattern;
 
 import static com.coolerpromc.uncrafteverything.config.UncraftEverythingConfig.tryParseTagKey;
@@ -97,7 +94,12 @@ public abstract class AbstractUncraftingTableBE extends BlockEntity {
 
         Tuple<List<UncraftingTableRecipe>, Boolean> outputs = UncraftingTableHelpers.getOutputs(inputStack, recipes, serverLevel);
         if (!outputs.getB()) return;
-        this.currentRecipes = outputs.getA();
+        if (UncraftEverythingConfig.CONFIG.prioritizeVanillaIngredientRecipe.get()){
+            this.currentRecipes = new ArrayList<>(outputs.getA().stream().sorted(Comparator.comparingInt(this::countVanillaIngredients).reversed()).toList());
+        }
+        else{
+            this.currentRecipes = outputs.getA();
+        }
 
         if (!currentRecipes.isEmpty()) {
             if (!isAuto){
@@ -122,6 +124,14 @@ public abstract class AbstractUncraftingTableBE extends BlockEntity {
                 this.status = Status.NO_RECIPE_FOUND;
             }
         }
+    }
+
+    public int countVanillaIngredients(UncraftingTableRecipe recipe){
+        int count = 0;
+        for (ItemStack stack : recipe.getOutputs()){
+            if (stack.getItemHolder().unwrapKey().get().location().getNamespace().equals("minecraft")) count++;
+        }
+        return count;
     }
 
     protected int getExperience(ItemStackHandler inputHandler) {