+++ src/main/java/com/coolerpromc/uncrafteverything/networking/ResponseConfigPayload.java
@@ -25,7 +25,8 @@ public record ResponseConfigPayload(
     Map<String, String> ftbQuestProgression,
     boolean enableProgression,
     boolean onlyAllowDefinedProgression,
-    boolean outputEnchantedBook
+    boolean outputEnchantedBook,
+    boolean prioritizeVanillaIngredientRecipe
 ) {
 
     public static final ResourceLocation TYPE = new ResourceLocation(UncraftEverything.MODID, "response_config");
@@ -57,6 +58,7 @@ public record ResponseConfigPayload(
         byteBuf.writeBoolean(payload.enableProgression);
         byteBuf.writeBoolean(payload.onlyAllowDefinedProgression);
         byteBuf.writeBoolean(payload.outputEnchantedBook);
+        byteBuf.writeBoolean(payload.prioritizeVanillaIngredientRecipe);
     }
 
     public static ResponseConfigPayload decode(FriendlyByteBuf byteBuf){
@@ -74,8 +76,9 @@ public record ResponseConfigPayload(
         boolean enableProgression = byteBuf.readBoolean();
         boolean onlyAllowDefinedProgression = byteBuf.readBoolean();
         boolean outputEnchantedBook = byteBuf.readBoolean();
+        boolean prioritizeVanillaIngredientRecipe = byteBuf.readBoolean();
 
-        return new ResponseConfigPayload(restrictionType, restrictedItems, allowEnchantedItem, experienceType, experience, allowUnsmithing, allowDamaged, preventModdedIngredientsFromVanillaItems, perItemExp, restrictedModIngredients, ftbQuestProgression, enableProgression, onlyAllowDefinedProgression, outputEnchantedBook);
+        return new ResponseConfigPayload(restrictionType, restrictedItems, allowEnchantedItem, experienceType, experience, allowUnsmithing, allowDamaged, preventModdedIngredientsFromVanillaItems, perItemExp, restrictedModIngredients, ftbQuestProgression, enableProgression, onlyAllowDefinedProgression, outputEnchantedBook, prioritizeVanillaIngredientRecipe);
     }
 
     public static void register(){