+++ src/main/java/com/coolerpromc/uncrafteverything/networking/ServerPayloadHandler.java
@@ -78,6 +78,7 @@ public class ServerPayloadHandler {
                 config.enableProgression.set(payload.enableProgression());
                 config.onlyAllowDefinedProgression.set(payload.onlyAllowDefinedProgression());
                 config.outputEnchantedBook.set(payload.outputEnchantedBook());
+                config.prioritizeVanillaIngredientRecipe.set(payload.prioritizeVanillaIngredientRecipe());
                 UncraftEverythingConfig.CONFIG_SPEC.save();
             }
         }).exceptionally(e -> {
@@ -104,7 +105,8 @@ public class ServerPayloadHandler {
                         FTBQuestProgressionConfig.getProgressionMap(),
                         config.enableProgression.get(),
                         config.onlyAllowDefinedProgression.get(),
-                        config.outputEnchantedBook.get()
+                        config.outputEnchantedBook.get(),
+                        config.prioritizeVanillaIngredientRecipe.get()
                 );
                 ResponseConfigPayload.INSTANCE.send(PacketDistributor.PLAYER.with(() -> context.get().getSender()), configPayload);
             }