+++ changelog.md
@@ -1,2 +1,1 @@
-# v2.0.1 - 1.20.1 Changelog
-- (Hopefully) Fixed uncrafted items unable to stack with crafted item
+- Added new common config field `prioritizeVanillaIngredientRecipe` to sort recipe selection order